Split e2e test
